refactor, added really bad let
